Asignación de funcionalidad
Se agregaron las llamadas a las nuevas interfaces
